requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.6,<=0.0.6
numpy>=1.23.5,<=1.23.5
opencv-python-headless>=4.7.0.72,<=4.7.0.72
requests>=2.28.1,<=2.28.1
imutils>=0.5.4,<=0.5.4
PyJWT==1.7.1
python-decouple==3.3
email-validator==1.1.2
mysqlclient==2.0.3
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
sqlalchemy==2.0.14
sendgrid==6.10.0
python-jose[cryptography]==3.3.0
